requirements.txt
pymongo==3.7.2
djangorestframework_simplejwt==4.4.0
Flask_JWT_Extended==3.9.1
pandas==1.1.5
Django==3.0.2
nltk==3.5
redis==3.3.11
spacy==2.3.5
mailer==0.8.1
numpy==1.19.1
djangorestframework==3.12.4
PyJWT==1.7.1
bcrypt==3.1.2
scikit_learn==0.24.2
django-sslserver==0.22
django-cors-headers==3.4.0
xlrd==2.0.1
openpyxl==3.0.7
XlsxWriter==1.4.3